requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pyodbc==5.1.0
pandas==2.2.0
plotly==5.19.0
python-dotenv==1.0.0